requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg==3.2.1
flask-jwt-extended==4.6.0
requests==2.32.4
sendgrid==6.11.0
twilio==9.3.0
python-dotenv==1.0.1
flask-cors==4.0.1
psycopg2==2.9.10
Pillow==11.3.0